main: Add navigation commands
Adds several commands to aid in navigating around a document. The
document now has a "focus" object that is the currently focused object.
Commands can use the "." shortcut to refer to the currently focused
object.

The "show" command now takes a path instead of just a handle. The path
can dereference object properties to show values instead of just relying
on the handle.

The "cd" command can be used to change the currently focused object,
using either a handle or a path.

The "rehandle" command can be used to change the handle of an object, in
case the user doesn't like the automatically assigned name.